backend/requirements.txt
fastapi
uvicorn[standard]       
python-dotenv
sqlalchemy
asyncpg                
geoalchemy2            
psycopg2       
requests
numpy
pandas
torch
scikit-learn
huggingface
transformers
geopandas
rasterio
sentinelsat
openeo
shapely
fiona
pyproj
geojson
emails
jinja2
celery
redis
pyyaml
email-validator